Split tests.rs into integration tests
